requirements.txt
Flask==2.3.2
nltk==3.8.1
spacy==3.5.1
es-core-news-sm==3.5.1
requests==2.31.0
python-telegram-bot==20.3
fpdf==1.7.2
pdf2image==1.16.3
pytesseract==0.3.10
